+++ src/test/java/address/JsonUtilTest.java
@@ -1,4 +1,4 @@
-package address.unittests;
+package address;
 
 import address.model.datatypes.AddressBook;
 import address.model.datatypes.ReadOnlyAddressBook;
