+++ src/test/java/address/sync/RemoteManagerTest.java
@@ -1,10 +1,7 @@
-package address.unittests.sync;
+package address.sync;
 
 import address.model.datatypes.person.Person;
 import address.model.datatypes.tag.Tag;
-import address.sync.ExtractedRemoteResponse;
-import address.sync.RemoteManager;
-import address.sync.RemoteService;
 import org.junit.Before;
 import org.junit.Test;
