+++ src/test/java/address/sync/cloud/CloudSimulatorTest.java
@@ -1,10 +1,6 @@
-package address.unittests.sync.cloud;
+package address.sync.cloud;
 
 import address.exceptions.DataConversionException;
-import address.sync.cloud.RemoteResponse;
-import address.sync.cloud.CloudFileHandler;
-import address.sync.cloud.CloudRateLimitStatus;
-import address.sync.cloud.CloudSimulator;
 import address.sync.cloud.model.CloudAddressBook;
 import address.sync.cloud.model.CloudPerson;
 import address.sync.cloud.model.CloudTag;
